fixup! admin: Add an `unlock` command, and libostree API
